# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBright Lights API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductCategory =
  (typeof ProductCategory)[keyof typeof ProductCategory];

export const ProductCategory = {
  LED_Lights: "LED Lights",
  Fans: "Fans",
  Solar_Products: "Solar Products",
} as const;

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  price: number;
  originalPrice?: number;
  description: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  badge?: string;
}

export type CreateProductCategory =
  (typeof CreateProductCategory)[keyof typeof CreateProductCategory];

export const CreateProductCategory = {
  LED_Lights: "LED Lights",
  Fans: "Fans",
  Solar_Products: "Solar Products",
} as const;

export interface CreateProduct {
  name: string;
  category: CreateProductCategory;
  price: number;
  originalPrice?: number;
  description: string;
  imageUrl: string;
  rating?: number;
  reviewCount?: number;
  inStock?: boolean;
  badge?: string;
}

export interface Review {
  id: number;
  customerName: string;
  rating: number;
  reviewText: string;
  createdAt: string;
}

export interface CreateReview {
  customerName: string;
  rating: number;
  reviewText: string;
}

export interface ContactForm {
  name: string;
  phone: string;
  message: string;
}

export interface ContactResponse {
  success: boolean;
  message: string;
}

export interface BusinessSettings {
  id: number;
  businessName: string;
  tagline: string;
  address: string;
  phone1: string;
  phone2: string;
  whatsapp: string;
  email: string;
  logoUrl: string;
  mapEmbed: string;
}

export interface UpdateBusinessSettings {
  businessName?: string;
  tagline?: string;
  address?: string;
  phone1?: string;
  phone2?: string;
  whatsapp?: string;
  email?: string;
  logoUrl?: string;
  mapEmbed?: string;
}

export interface UploadResponse {
  url: string;
  filename: string;
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export type GetProductsParams = {
  category?: string;
};

export type AdminUploadFileBody = {
  file: Blob;
};