# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EcoBright Lights API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product operations
  - name: reviews
    description: Review operations
  - name: contact
    description: Contact form operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get all reviews
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReview"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactForm"
      responses:
        "201":
          description: Contact form submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"

  /admin/settings:
    get:
      operationId: getBusinessSettings
      tags: [admin]
      summary: Get business settings
      responses:
        "200":
          description: Business settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"
    put:
      operationId: updateBusinessSettings
      tags: [admin]
      summary: Update business settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"

  /admin/products:
    get:
      operationId: adminGetProducts
      tags: [admin]
      summary: Get all products (admin)
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"

  /admin/upload:
    post:
      operationId: adminUploadFile
      tags: [admin]
      summary: Upload an image file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [LED Lights, Fans, Solar Products]
        price:
          type: number
        originalPrice:
          type: number
        description:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        badge:
          type: string
      required:
        - id
        - name
        - category
        - price
        - description
        - imageUrl
        - rating
        - reviewCount
        - inStock

    CreateProduct:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [LED Lights, Fans, Solar Products]
        price:
          type: number
        originalPrice:
          type: number
        description:
          type: string
        imageUrl:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        inStock:
          type: boolean
        badge:
          type: string
      required:
        - name
        - category
        - price
        - description
        - imageUrl

    Review:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        reviewText:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerName
        - rating
        - reviewText
        - createdAt

    CreateReview:
      type: object
      properties:
        customerName:
          type: string
        rating:
          type: integer
        reviewText:
          type: string
      required:
        - customerName
        - rating
        - reviewText

    ContactForm:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        message:
          type: string
      required:
        - name
        - phone
        - message

    ContactResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    BusinessSettings:
      type: object
      properties:
        id:
          type: integer
        businessName:
          type: string
        tagline:
          type: string
        address:
          type: string
        phone1:
          type: string
        phone2:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        logoUrl:
          type: string
        mapEmbed:
          type: string
      required:
        - id
        - businessName
        - tagline
        - address
        - phone1
        - phone2
        - whatsapp
        - email
        - logoUrl
        - mapEmbed

    UpdateBusinessSettings:
      type: object
      properties:
        businessName:
          type: string
        tagline:
          type: string
        address:
          type: string
        phone1:
          type: string
        phone2:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        logoUrl:
          type: string
        mapEmbed:
          type: string

    UploadResponse:
      type: object
      properties:
        url:
          type: string
        filename:
          type: string
      required:
        - url
        - filename

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message