# lib/api-zod/src/generated/types/createProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBright Lights API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProductCategory } from "./createProductCategory";

export interface CreateProduct {
  name: string;
  category: CreateProductCategory;
  price: number;
  originalPrice?: number;
  description: string;
  imageUrl: string;
  rating?: number;
  reviewCount?: number;
  inStock?: boolean;
  badge?: string;
}