# lib/api-zod/src/generated/types/updateBusinessSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EcoBright Lights API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBusinessSettings {
  businessName?: string;
  tagline?: string;
  address?: string;
  phone1?: string;
  phone2?: string;
  whatsapp?: string;
  email?: string;
  logoUrl?: string;
  mapEmbed?: string;
}